+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.response;
+package cart.dto.response;
 
 import java.util.Objects;
