+++ src/main/java/cart/service/ProductService.java
@@ -1,8 +1,8 @@
 package cart.service;
 
 import cart.dao.ProductDao;
-import cart.domain.Product;
-import cart.dto.ProductDto;
+import cart.domain.product.Product;
+import cart.dto.application.ProductDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
