+++ src/test/java/cart/controller/ProductResourceControllerTest.java
@@ -1,6 +1,6 @@
 package cart.controller;
 
-import cart.request.ProductRequest;
+import cart.dto.request.ProductRequest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
